backend/requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
flatbuffers==24.3.7
fsspec==2024.3.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.3
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.17.1
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
safetensors==0.4.2
sqlparse==0.4.4
sympy==1.12
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.39.1
typing_extensions==4.9.0
urllib3==2.2.1
